+++ .config/dotnet-tools.json
@@ -1,0 +1,12 @@
+{
+  "version": 1,
+  "isRoot": true,
+  "tools": {
+    "microsoft.tye": {
+      "version": "0.11.0-alpha.22112.1",
+      "commands": [
+        "tye"
+      ]
+    }
+  }
+}
