+++ Udap.Server/Registration/IUdapDynamicClientRegistrationValidator.cs
@@ -9,6 +9,7 @@
 
 using System.Security.Cryptography.X509Certificates;
 using Udap.Client.Client.Messages;
+using Udap.Model.Registration;
 
 namespace Udap.Server.Registration;
