--- examples/clients/blazor7/UdapClient/Client/Pages/Counter.razor
@@ -1,20 +1,0 @@
-﻿@page "/counter"
-
-<MudGrid Class="mt-2">
-  <MudItem xs="12" sm="12" md="12">
-    <MudText Typo="Typo.h3">Counter</MudText>
-    <MudText Typo="Typo.subtitle1">Current count: @_currentCount</MudText>
-  </MudItem>
-  <MudItem xs="12" sm="12" md="12">
-    <MudButton Variant="Variant.Filled" Color="Color.Primary" OnClick="IncrementCount">Click me</MudButton>
-  </MudItem>
-</MudGrid>
-
-@code {
-    private int _currentCount;
-
-    private void IncrementCount()
-    {
-        _currentCount++;
-    }
-}
