+++ examples/clients/blazor7/UdapClient/Client/Pages/Index.razor
@@ -2,10 +2,16 @@
 
 <MudGrid Class="mt-4">
     <MudItem xs="12" sm="12" md="12">
-        <MudText Typo="Typo.h3">Hello, world!</MudText>
-        <MudText Typo="Typo.subtitle1">Welcome to your new app.</MudText>
+        <MudText Typo="Typo.h3">UDAP Client Lab</MudText>
+        <MudSpacer />
+        <MudText Typo="Typo.subtitle2">Experiment with:</MudText>
+    <MudSpacer />
+    <ul style="list-style:disc">
+            <li>Metadata discovery</li>
+            <li>Software statement</li>
+            <li>Registration request for client_credential and authorization_code flow</li>
+            <li>Certificate trust validation</li>
+        </ul>
     </MudItem>
-    <MudItem xs="12" sm="12" md="12">
-        <SurveyPrompt Title="How is Blazor working for you?" />
-    </MudItem>
+    
 </MudGrid>
