+++ examples/clients/blazor7/UdapClient/Client/Pages/UdapDiscovery.razor.cs
@@ -1,4 +1,13 @@
-﻿using Microsoft.AspNetCore.Components;
+﻿#region (c) 2023 Joseph Shook. All rights reserved.
+// /*
+//  Authors:
+//     Joseph Shook   Joseph.Shook@Surescripts.com
+// 
+//  See LICENSE in the project root for license information.
+// */
+#endregion
+
+using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Web;
 using Udap.Model;
 using UdapClient.Client.Services;
