+++ examples/clients/blazor7/UdapClient/Client/Pages/UdapRegistration.razor
@@ -4,11 +4,12 @@
 
 <ErrorBoundary @ref="ErrorBoundary">
   <ChildContent>
-    <MudGrid Class="mt-4">
+
+<MudGrid Class="mt-4">
       <MudItem xs="12" sm="12" md="12">
         <MudText Typo="Typo.h3">Udap Metadata</MudText>
       </MudItem>
-        <MudItem>
+      <MudItem xs="12" sm="12" md="12">
             <MudFab HtmlTag="label"
                     Size="Size.Small"
                     Color="Color.Success"
