+++ examples/clients/blazor7/UdapClient/Client/Pages/UdapRegistration.razor.cs
@@ -1,18 +1,17 @@
-﻿using System.IdentityModel.Tokens.Jwt;
-using System.Security.Cryptography.X509Certificates;
-using System.Text;
-using System.Text.Encodings.Web;
-using System.Text.Json.Nodes;
-using System.Text.Json;
-using IdentityModel;
+﻿#region (c) 2023 Joseph Shook. All rights reserved.
+// /*
+//  Authors:
+//     Joseph Shook   Joseph.Shook@Surescripts.com
+// 
+//  See LICENSE in the project root for license information.
+// */
+#endregion
+
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Forms;
 using Microsoft.AspNetCore.Components.Web;
-using Microsoft.IdentityModel.JsonWebTokens;
-using Microsoft.IdentityModel.Tokens;
-using Udap.Model;
-using Udap.Model.Registration;
 using UdapClient.Client.Services;
+using UdapClient.Shared.Model;
 
 namespace UdapClient.Client.Pages;
 
@@ -30,59 +29,14 @@ public partial class UdapRegistration
     {
         try
         {
-            var now = DateTime.UtcNow;
-            var jwtId = CryptoRandom.CreateUniqueId();
-
-            var clientCert = new X509Certificate2(UdapClientState.ClientCert);
-            var securityKey = new X509SecurityKey(clientCert);
-            var signingCredentials = new SigningCredentials(securityKey, UdapConstants.SupportedAlgorithm.RS256);
-
-            var certBase64 = Convert.ToBase64String(clientCert.Export(X509ContentType.Cert));
-            var jwtHeader = new JwtHeader
-            {
-                { "alg", signingCredentials.Algorithm },
-                { "x5c", new[] { certBase64 } }
-            };
-
-            var document = new UdapDynamicClientRegistrationDocument
-            {
-                Issuer = "https://fhirlabs.net:7016/fhir/r4",
-                Subject = "https://fhirlabs.net:7016/fhir/r4",
-                Audience = UdapClientState.MetadataUrl,
-                Expiration = EpochTime.GetIntDate(now.AddMinutes(5).ToUniversalTime()),
-                IssuedAt = EpochTime.GetIntDate(now.ToUniversalTime()),
-                JwtId = jwtId,
-                ClientName = "udapTestClient",
-                Contacts = new HashSet<string> { "mailto:Joseph.Shook@Surescripts.com", "mailto:JoeShook@gmail.com" },
-                GrantTypes = new HashSet<string> { "client_credentials" },
-                TokenEndpointAuthMethod = UdapConstants.RegistrationDocumentValues.TokenEndpointAuthMethodValue,
-                Scope = "system/Patient.* system/Practitioner.read"
-            };
-
-            var encodedHeader = jwtHeader.Base64UrlEncode();
-            var encodedPayload = document.Base64UrlEncode();
-            var encodedSignature =
-                JwtTokenUtilities.CreateEncodedSignature(string.Concat(encodedHeader, ".", encodedPayload),
-                    signingCredentials);
-            var signedSoftwareStatement = string.Concat(encodedHeader, ".", encodedPayload, ".", encodedSignature);
+            var request = new BuildSoftwareStatementRequest();
+            request.MetadataUrl = UdapClientState.MetadataUrl;
             
-            var tokenHandler = new JsonWebTokenHandler();
-            var jsonToken = tokenHandler.ReadToken(signedSoftwareStatement);
-            var requestToken = jsonToken as JsonWebToken;
+            //TODO Get from User:: Dialog or form
+            request.Password = "udap-test";
 
-            var sb = new StringBuilder();
-            sb.Append("[");
-            sb.Append(Base64UrlEncoder.Decode(requestToken.EncodedHeader));
-            sb.Append(",");
-            sb.Append(Base64UrlEncoder.Decode(requestToken.EncodedPayload));
-            sb.Append("]");
-
-            SoftwareStatementBeforeEncoding = JsonObject.Parse(sb?.ToString())
-                .ToJsonString(new JsonSerializerOptions()
-                {
-                    WriteIndented = true, 
-                    Encoder = JavaScriptEncoder.UnsafeRelaxedJsonEscaping
-                });
+            SoftwareStatementBeforeEncoding = await MetadataService.BuildSoftwareStatement(request);
+            UdapClientState.SoftwareStatementBeforeEncoding = SoftwareStatementBeforeEncoding;
         }
         catch (Exception ex)
         {
@@ -90,11 +44,14 @@ public partial class UdapRegistration
         }
     }
 
-    private void BuildRequestBody()
+    private async Task BuildRequestBody()
     {
-        RequestBody = string.Empty;
-    }
+        var request = new BuildSoftwareStatementRequest();
+        request.MetadataUrl = UdapClientState.MetadataUrl;
+        request.Password = "udap-test";
 
+        RequestBody = await MetadataService.BuildRequestBody(request);
+    }
     
 
     private async Task UploadFilesAsync(InputFileChangeEventArgs e)
