+++ examples/clients/blazor7/UdapClient/Client/Services/ProfileService.cs
@@ -1,4 +1,13 @@
-﻿using Blazored.LocalStorage;
+﻿#region (c) 2023 Joseph Shook. All rights reserved.
+// /*
+//  Authors:
+//     Joseph Shook   Joseph.Shook@Surescripts.com
+// 
+//  See LICENSE in the project root for license information.
+// */
+#endregion
+
+using Blazored.LocalStorage;
 
 namespace UdapClient.Client.Services;
