+++ examples/clients/blazor7/UdapClient/Client/Shared/NavMenu.razor
@@ -2,8 +2,6 @@
   <MudNavLink Href="" Match="NavLinkMatch.All" Icon="@Icons.Filled.Dashboard" IconColor="Color.Inherit">
     Home
   </MudNavLink>
-  <MudNavLink Href="counter" Icon="@Icons.Filled.Timer" IconColor="Color.Inherit">Counter</MudNavLink>
-    <MudNavLink Href="fetchdata" Icon="@Icons.Filled.FileDownload" IconColor="Color.Inherit">Weather data</MudNavLink>
     <MudNavLink Href="udapDiscovery" Icon="@Icons.Filled.Dataset" IconColor="Color.Inherit">Discovery</MudNavLink>
     <MudNavLink Href="udapRegistration" Icon="@Icons.Filled.Dataset" IconColor="Color.Inherit">Registration</MudNavLink>
 </MudNavMenu>
