+++ examples/clients/blazor7/UdapClient/Shared/JsonExtensions.cs
@@ -1,4 +1,13 @@
-﻿using System.Text.Json;
+﻿#region (c) 2023 Joseph Shook. All rights reserved.
+// /*
+//  Authors:
+//     Joseph Shook   Joseph.Shook@Surescripts.com
+// 
+//  See LICENSE in the project root for license information.
+// */
+#endregion
+
+using System.Text.Json;
 
 namespace UdapClient.Shared;
 public static class JsonExtensions
