--- examples/clients/blazor7/UdapClient/Shared/WeatherForecast.cs
@@ -1,12 +1,0 @@
-namespace UdapClient.Shared;
-
-public class WeatherForecast
-{
-    public DateOnly Date { get; set; }
-
-    public int TemperatureC { get; set; }
-
-    public string? Summary { get; set; }
-
-    public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
-}
