+++ examples/clients/blazor7/mudblazor/mudblazor.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.4.33205.214
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JoeMudBlazor", "JoeMudBlazor.csproj", "{EAB6517B-F877-4D01-887E-C3629C515951}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{EAB6517B-F877-4D01-887E-C3629C515951}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EAB6517B-F877-4D01-887E-C3629C515951}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EAB6517B-F877-4D01-887E-C3629C515951}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EAB6517B-F877-4D01-887E-C3629C515951}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {4931BAFE-8CA5-4BA5-BD41-ECBB6AE03BE0}
+	EndGlobalSection
+EndGlobal
